requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
